kiddie_closet_app/serializers/neighborhoods.py
from rest_framework import serializers

from kiddie_closet_app.models import Neighborhood


class NeighborhoodSerializer(serializers.ModelSerializer):
    class Meta:
        model = Neighborhood
        fields = "__all__"